requirements.txt
streamlit
jinja2
python-docx
weasyprint
python-dotenv
streamlit
bs4
fpdf
fastapi
uvicorn
